+++ Realm/ObjectServerTests/RealmServer.swift
@@ -80,8 +80,6 @@ private func bsonType(_ type: PropertyType) -> String {
     case .objectId: return "objectId"
     case .string: return "string"
     case .linkingObjects: return "linkingObjects"
-    default: // Types must be added when tested for sync
-        fatalError()
     }
 }