+++ Realm/RLMConstants.h
@@ -95,43 +95,43 @@ typedef RLM_CLOSED_ENUM(int32_t, RLMPropertyType) {
     RLMPropertyTypeLinkingObjects = 8,
 };
 
-typedef RLM_CLOSED_ENUM(int32_t, RLMMixedValueType) {
+typedef RLM_CLOSED_ENUM(int32_t, RLMAnyValueType) {
 #pragma mark - Primitive types
     /** Integers: `NSInteger`, `int`, `long`, `Int` (Swift) */
-    RLMMixedValueTypeInt    = 0,
+    RLMAnyValueTypeInt    = 0,
     /** Booleans: `BOOL`, `bool`, `Bool` (Swift) */
-    RLMMixedValueTypeBool   = 1,
+    RLMAnyValueTypeBool   = 1,
     /** Floating-point numbers: `float`, `Float` (Swift) */
-    RLMMixedValueTypeFloat  = 5,
+    RLMAnyValueTypeFloat  = 5,
     /** Double-precision floating-point numbers: `double`, `Double` (Swift) */
-    RLMMixedValueTypeDouble = 6,
+    RLMAnyValueTypeDouble = 6,
     /** NSUUID, UUID */
-    RLMMixedValueTypeUUID   = 12,
+    RLMAnyValueTypeUUID   = 12,
 
 #pragma mark - Object types
 
     /** Strings: `NSString`, `String` (Swift) */
-    RLMMixedValueTypeString = 2,
+    RLMAnyValueTypeString = 2,
     /** Binary data: `NSData` */
-    RLMMixedValueTypeData   = 3,
+    RLMAnyValueTypeData   = 3,
     /** Any type: `id<RLMValue>`, `AnyRealmValue` (Swift) */
-    RLMMixedValueTypeAny    = 9,
+    RLMAnyValueTypeAny    = 9,
     /** Dates: `NSDate` */
-    RLMMixedValueTypeDate   = 4,
-    RLMMixedValueTypeObjectId = 10,
-    RLMMixedValueTypeDecimal128 = 11,
+    RLMAnyValueTypeDate   = 4,
+    RLMAnyValueTypeObjectId = 10,
+    RLMAnyValueTypeDecimal128 = 11,
 
 #pragma mark - Linked object types
 
     /** Realm model objects. See [Realm Models](https://www.mongodb.com/docs/realm/sdk/swift/fundamentals/object-models-and-schemas/) for more information. */
-    RLMMixedValueTypeObject = 7,
+    RLMAnyValueTypeObject = 7,
     /** Realm linking objects. See [Realm Models](https://www.mongodb.com/docs/realm/sdk/swift/fundamentals/relationships/#inverse-relationship) for more information. */
-    RLMMixedValueTypeLinkingObjects = 8,
+    RLMAnyValueTypeLinkingObjects = 8,
 
     /** Dictionary: `RLMDictionary`, `Map` (Swift) */
-    RLMMixedValueTypeDictionary = 512,
+    RLMAnyValueTypeDictionary = 512,
     /** Set: `RLMArray`, `List` (Swift) */
-    RLMMixedValueTypeList = 128,
+    RLMAnyValueTypeList = 128,
 };
 
 #pragma mark - Notification Constants