+++ Realm/RLMValue.h
@@ -62,7 +62,9 @@
 @protocol RLMValue
 
 /// Describes the type of property stored.
-@property (readonly) RLMMixedValueType rlm_valueType;
+@property (readonly) RLMAnyValueType rlm_valueType __attribute__((deprecated("Use `rlm_anyValueType` instead, which includes collection types as well")));
+/// Describes the type of property stored.
+@property (readonly) RLMAnyValueType rlm_anyValueType;
 
 @end